backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "bcrypt==3.2.2",
    "fastapi[standard]>=0.124.4",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]
